Restructured the card presenters and views to fit the model.view-presenter standard
