+++ src/presenters/boardPresenter.js
@@ -1,10 +1,9 @@
 import React from "react";
-import { CardView, BoardView } from "../views";
+import { BoardView } from "../views";
 
 import useBoardProp from "./useBoardProp";
-import useCardProp from "./useCardProp";
 
-import CardPresenter from "./translationCardPresenter";
+import CardPresenter from "./cardPresenter";
 
 /*
     BoardPresenter 
@@ -13,10 +12,7 @@ import CardPresenter from "./translationCardPresenter";
 */
 
 export default function BoardPresenter(props) {
-  // Using boardIndex to accses the right board element in the model
-  // const board = useModelSubSubclassProperty(props.model, "banks", props.model.activeBankId, "boards", props.boardIndex);
-  // const board = useBoardProp(props.model, "banks", props.model.activeBankId, "boards", props.boardIndex);
-  const cards = useBoardProp(props.model, props.id, "cards");
+    const cards = useBoardProp(props.model, props.id, "cards");
 
   var cardPresenterList = cards.map(function (card, key) {
     return (
