+++ src/views/translationCardView.js
@@ -14,13 +14,12 @@ import {
 } from "../styledComponents";
 
 export default function TranslationCardView(props) {
+
   return (
     <Card>
       <span
         key={props.id}
-        onClick={() => {
-          props.onCardPress();
-        }}
+        onClick={() => { props.onCardPress() }}
       >
         <TranslationWrapper>
           <Translation isTranslateFrom={true}>
